backend/requirement.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
google-generativeai==0.3.0
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6
